refactor(threads): merge useTaskActions into useThreadActions; drop legacy KEYS.tasks
Migrate 4 useTaskActions callers to useThreadActions (absorbs the full low-level
CRUD surface). Replace tasksPrefix cache invalidations with targeted prepends
(for creates) or bridge-handles comments (for updates). Remove legacy tasks() and
tasksPrefix() keys from query-keys.ts. Delete use-tasks.ts hook. Update test
assertions to use threadsPrefix. Fix unused variable in use-task-expanded-tools.
