+++ apps/mesh/src/web/layouts/shell-layout.tsx
@@ -17,7 +17,7 @@ import {
 } from "@tanstack/react-router";
 import { KEYS } from "../lib/query-keys";
 import { readCachedTaskBranch } from "../lib/read-cached-task-branch";
-import { useTaskActions } from "../hooks/use-tasks";
+import { useThreadActions } from "@/web/components/chat/task";
 import { useOrganizationSettingsSuspense } from "../hooks/use-organization-settings";
 import { useOrgSsoStatus } from "../hooks/use-org-sso";
 import { SsoRequiredScreen } from "../components/sso-required-screen";
@@ -67,7 +67,7 @@ function ShellProjectProvider({
 export function usePanelActions() {
   const navigate = useNavigate();
   const queryClient = useQueryClient();
-  const taskActions = useTaskActions();
+  const taskActions = useThreadActions();
   const { locator } = useProjectContext();
 
   const params = useParams({ strict: false }) as {
